Factor out complexity
