+++ src/distributed_planner/statistics/complexity_cpu.rs
@@ -1,7 +1,8 @@
 use crate::BroadcastExec;
+use crate::distributed_planner::statistics::complexity::{Complexity, LinearComplexity};
 use crate::execution_plans::ChildrenIsolatorUnionExec;
 use datafusion::catalog::memory::DataSourceExec;
-use datafusion::common::{JoinSide, Statistics};
+use datafusion::common::JoinSide;
 use datafusion::physical_expr::PhysicalExpr;
 use datafusion::physical_plan::aggregates::AggregateExec;
 use datafusion::physical_plan::coalesce_partitions::CoalescePartitionsExec;
@@ -20,166 +21,10 @@ use datafusion::physical_plan::sorts::sort_preserving_merge::SortPreservingMerge
 use datafusion::physical_plan::union::{InterleaveExec, UnionExec};
 use datafusion::physical_plan::windows::{BoundedWindowAggExec, WindowAggExec};
 use datafusion::physical_plan::{ExecutionPlan, Partitioning};
-use std::fmt::{Debug, Formatter};
 use std::sync::Arc;
 
-#[derive(Clone, PartialEq, Eq)]
-pub(super) enum Complexity {
-    /// Constant complexity
-    Constant(usize),
-    /// Linear with a specific column from a specific child.
-    Linear(LinearComplexity),
-    /// NLogM
-    Log(Box<Complexity>, Box<Complexity>),
-    /// N+M
-    Plus(Box<Complexity>, Box<Complexity>),
-    /// N*M
-    Multiply(Box<Complexity>, Box<Complexity>),
-}
-
-#[derive(Clone, PartialEq, Eq)]
-pub(super) enum LinearComplexity {
-    /// Depends on linearly with the input column with the provided index
-    Column(usize),
-    /// Depends on linearly with the all the input columns
-    AllColumns,
-    /// Depends on linearly with the input column with the provided index from the left child
-    ColumnFromLeft(usize),
-    /// Depends on linearly with the all the input columns from the left child
-    AllColumnsFromLeft,
-    /// Depends on linearly with the input column with the provided index from the right child
-    ColumnFromRight(usize),
-    /// Depends on linearly with the all the input columns from the right child
-    AllColumnsFromRight,
-    /// Depends on linearly with the all the output columns
-    AllOutputColumns,
-}
-
-impl Complexity {
-    fn log(self, other: Self) -> Self {
-        match (self, other) {
-            (Self::Constant(n), Self::Constant(m)) => {
-                Self::Constant(n * (m as f64).log2() as usize)
-            }
-            (s, o) => Self::Log(Box::new(s), Box::new(o)),
-        }
-    }
-
-    fn plus(self, other: Self) -> Self {
-        match (self, other) {
-            (Self::Constant(n), Self::Constant(m)) => Self::Constant(n + m),
-            // (A + k1) + k2 = A + (k1 + k2): bubble constants rightward so they can fold
-            (Self::Plus(a, b), Self::Constant(m)) if matches!(*b, Self::Constant(_)) => {
-                (*a).plus((*b).plus(Self::Constant(m)))
-            }
-            (s, o) if s == o => Self::Constant(2).multiply(s),
-            (s, o) => Self::Plus(Box::new(s), Box::new(o)),
-        }
-    }
-
-    fn multiply(self, other: Self) -> Self {
-        match (self, other) {
-            (Self::Constant(n), Self::Constant(m)) => Self::Constant(n * m),
-            (s, o) => Self::Multiply(Box::new(s), Box::new(o)),
-        }
-    }
-
-    /// Computes the total bytes processed given per-child row counts.
-    /// Returns None if statistics are unavailable for any required input.
-    pub(super) fn cost(
-        &self,
-        output_stat: &Arc<Statistics>,
-        input_stats: &[Arc<Statistics>],
-    ) -> Option<usize> {
-        Some(match self {
-            Self::Constant(v) => *v,
-            Self::Linear(linear) => match linear {
-                LinearComplexity::Column(i) => {
-                    let col_stats = &input_stats.first()?.column_statistics;
-                    *col_stats.get(*i)?.byte_size.get_value()?
-                }
-                LinearComplexity::AllColumns => {
-                    *input_stats.first()?.total_byte_size.get_value()?
-                }
-                LinearComplexity::ColumnFromLeft(i) => {
-                    let col_stats = &input_stats.first()?.column_statistics;
-                    *col_stats.get(*i)?.byte_size.get_value()?
-                }
-                LinearComplexity::AllColumnsFromLeft => {
-                    *input_stats.first()?.total_byte_size.get_value()?
-                }
-                LinearComplexity::ColumnFromRight(i) => {
-                    let col_stats = &input_stats.last()?.column_statistics;
-                    *col_stats.get(*i)?.byte_size.get_value()?
-                }
-                LinearComplexity::AllColumnsFromRight => {
-                    *input_stats.last()?.total_byte_size.get_value()?
-                }
-                LinearComplexity::AllOutputColumns => *output_stat.total_byte_size.get_value()?,
-            },
-            Self::Log(n, m) => {
-                let n = n.cost(output_stat, input_stats)?;
-                let m = m.cost(output_stat, input_stats)?;
-                // `ilog2` panics on 0, which happens whenever the logged input has zero estimated
-                // bytes/rows (e.g. an empty or fully-pruned relation). Flooring at 1 makes log2
-                // contribute 0 there, i.e. sorting/merging nothing costs nothing.
-                n * m.checked_ilog2().unwrap_or(0) as usize
-            }
-            Self::Plus(n, m) => n
-                .cost(output_stat, input_stats)?
-                .saturating_add(m.cost(output_stat, input_stats)?),
-            Self::Multiply(n, m) => n
-                .cost(output_stat, input_stats)?
-                .saturating_mul(m.cost(output_stat, input_stats)?),
-        })
-    }
-}
-
-impl Debug for Complexity {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        fn trim_parenthesis(dbg: &Complexity) -> String {
-            let s = format!("{dbg:?}");
-            if s.starts_with('(') && s.ends_with(')') {
-                s[1..s.len() - 1].to_string()
-            } else {
-                s
-            }
-        }
-        match self {
-            Self::Constant(v) => write!(f, "{v}"),
-            Self::Linear(linear) => match linear {
-                LinearComplexity::Column(i) => write!(f, "Col{i}"),
-                LinearComplexity::AllColumns => write!(f, "Cols"),
-                LinearComplexity::ColumnFromLeft(i) => write!(f, "left_Col{i}"),
-                LinearComplexity::AllColumnsFromLeft => write!(f, "left_Cols"),
-                LinearComplexity::ColumnFromRight(i) => write!(f, "right_Col{i}"),
-                LinearComplexity::AllColumnsFromRight => write!(f, "right_Cols"),
-                LinearComplexity::AllOutputColumns => write!(f, "out_Cols"),
-            },
-            Self::Log(n, m) => write!(f, "{n:?}*Log({m:?})"),
-            Self::Plus(n, m) => {
-                if matches!(n.as_ref(), &Self::Plus(_, _)) {
-                    write!(f, "({}+{m:?})", trim_parenthesis(n))
-                } else {
-                    write!(f, "({n:?}+{m:?})")
-                }
-            }
-            Self::Multiply(n, m) => {
-                if matches!(n.as_ref(), &Self::Multiply(_, _)) {
-                    write!(f, "({}*{m:?})", trim_parenthesis(n))
-                } else {
-                    write!(f, "({n:?}*{m:?})")
-                }
-            }
-        }
-    }
-}
-
-/// Calculates what's the cost, expressed as a number, per input row for each input children.
-///
-/// The Vec return has equal size to `node.children()`, and determines how many each input needs
-/// to be processed
-pub(super) fn calculate_compute_complexity(node: &Arc<dyn ExecutionPlan>) -> Complexity {
+/// Calculates the CPU cost for the provided node, without recursing into children.
+pub(super) fn complexity_cpu(node: &Arc<dyn ExecutionPlan>) -> Complexity {
     // NestedLoopJoinExec: O(n*m) - evaluates join condition for each pair of rows
     // https://github.com/apache/datafusion/blob/branch-52/datafusion/physical-plan/src/joins/nested_loop_join.rs
     if let Some(node) = node.downcast_ref::<NestedLoopJoinExec>() {
@@ -582,7 +427,7 @@ fn _expression_complexity(expression: &Arc<dyn PhysicalExpr>) -> BytesPerRow {
 #[cfg(test)]
 mod tests {
     use crate::assert_snapshot;
-    use crate::distributed_planner::statistics::compute_per_node::calculate_compute_complexity;
+    use crate::distributed_planner::statistics::complexity_cpu::complexity_cpu;
     use crate::test_utils::plans::TestPlanBuilder;
     use datafusion::common::tree_node::{Transformed, TreeNode};
     use datafusion::physical_plan::{ExecutionPlan, displayable};
@@ -1012,7 +857,7 @@ mod tests {
                 let node = displayable(plan.as_ref()).one_line().to_string();
                 display += &format!(
                     "{indent}O({:?}) | {}\n",
-                    calculate_compute_complexity(&plan),
+                    complexity_cpu(&plan),
                     node.trim_end()
                 );
                 *depth.borrow_mut() += 1;
