+++ src/distributed_planner/statistics/mod.rs
@@ -1,4 +1,5 @@
-mod compute_per_node;
+mod complexity;
+mod complexity_cpu;
 mod cost;
 mod default_bytes_for_datatype;
 mod plan_statistics;
